requirements.txt
python-dotenv==1.0.1
RPi.GPIO==0.7.1
